// ro-sec-fe-angular-dev/Source/src/app/interfaces/transaction.ts
export interface Transaction {
    id: string;
    // firstName: string;
    // lastName: string;
    // cnp: string;
    // seria: string;
    // number: string;
    // gender: string;
    // dob: string;
    // supplier: string;
    // subscription: string;
    // obs: string;
    // history: string;

    clientId: string;
    bookId: string;
    price: string;
    rentalDate: string;
    returnDate: string;
    rentalDuration: string;
}