// ro-sec-fe-angular-dev/Source/src/app/interfaces/clientResponse.ts
export interface IClientResponse{
    id: string;
    // firstName: string;
    // lastName: string;
    // cnp: string;
    // seria: string;
    // number: string;
    // gender: string;
    // dob: string;
    // supplier: string;
    // subscription: string;
    // obs: string;
    // history: string;
    // error: string[];
    // logError: boolean;

    firstName: string;
    lastName: string;
    email: string;
    error: string[];
    logError: boolean;

}