// ro-sec-fe-angular-dev/Source/src/app/interfaces/bookResponse.ts
export interface IBookResponse{
    id: string;
    // firstName: string;
    // lastName: string;
    // cnp: string;
    // seria: string;
    // number: string;
    // gender: string;
    // dob: string;
    // supplier: string;
    // subscription: string;
    // obs: string;
    // history: string;
    // error: string[];
    // logError: boolean;

    title: string;
    authorId: string;
    publicationDate: string;
    quantity: string;
    error: string[];
    logError: boolean;

}